Moved JSON files to their own directory
