+++ flexy-osisoft-connector/src/com/hms/flexyosisoftconnector/JSON/JSONString.java
@@ -1,4 +1,4 @@
-package com.hms.flexyosisoftconnector;
+package com.hms.flexyosisoftconnector.JSON;
 
 /**
  * The <code>JSONString</code> interface allows a <code>toJSONString()</code>
