+++ flexy-osisoft-connector/src/com/hms/flexyosisoftconnector/JSON/StringWriter.java
@@ -1,4 +1,4 @@
-package com.hms.flexyosisoftconnector;
+package com.hms.flexyosisoftconnector.JSON;
 /*
 Copyright (c) 2002 JSON.org
