+++ flexy-osisoft-connector/src/com/hms/flexyosisoftconnector/OSIsoftConfig.java
@@ -1,5 +1,6 @@
 package com.hms.flexyosisoftconnector;
 import java.util.ArrayList;
+import com.hms.flexyosisoftconnector.JSON.*;
 
 /**
  *
